Add support for fstype in default partitioning config
Allow setting specific fstype for individual partitions in the config.
This specifically allows different fstypes for different partitions.
The feature is optional - if fstype is not set, the behavior is
unchanged.

v2: drop unit test from AutomaticPartitioningTaskTestCase, as it isn't
part of the release tarball so the patch fails to apply