Set 512 LUKS sector size under LVM
LVM (thin) volumes present the original sector size to the VM. Currently
all the qubes templates are built with 512 sector size (that's how
partition table and filesystem are created). When installing Qubes on a
4Kn disk, such VMs won't boot. LUKS2 supports passing down the sector
size, and cryptsetup used in Qubes 4.2 has it enabled by default.

As a temporary workaround, force 512 LUKS sector size on a LVM thin
partitioning layout.

QubesOS/qubes-issues#4974